영화 리뷰 정보 추가 (#123)
* 영화 리뷰 정보 추가

* 테스트 수정

* 테스트 수정

* 테스트 수정

* 테스트 수정

+++ core/model/src/main/java/com/bowoon/model/MovieReviews.kt
@@ -31,6 +31,6 @@ data class MovieReview(
 data class MovieReviewAuthorDetails(
     val avatarPath: String? = null,
     val name: String? = null,
-    val rating: Int? = null,
+    val rating: Float? = null,
     val username: String? = null
 ) : Parcelable
